requirements.txt
haversine
meshtastic
requests
pyyaml
